+++ packages/@ourworldindata/grapher/src/scatterCharts/ScatterPlotChart.tsx
@@ -62,6 +62,7 @@ import {
     ConnectedScatterLegendManager,
 } from "./ConnectedScatterLegend"
 import { VerticalColorLegend } from "../verticalColorLegend/VerticalColorLegend"
+import { VerticalColorLegendComponent } from "../verticalColorLegend/VerticalColorLegendComponent"
 import { DualAxisComponent } from "../axis/AxisViews"
 import { DualAxis, HorizontalAxis, VerticalAxis } from "../axis/Axis"
 
@@ -506,11 +507,8 @@ export class ScatterPlotChart
         return this.tooltipState.target?.series
     }
 
-    @computed private get verticalColorLegend(): {
-        width: number
-        height: number
-    } {
-        return VerticalColorLegend.dimensions({
+    @computed private get verticalColorLegend(): VerticalColorLegend {
+        return new VerticalColorLegend({
             maxLegendWidth: this.maxLegendWidth,
             fontSize: this.fontSize,
             legendItems: this.legendItems,
@@ -832,7 +830,16 @@ export class ScatterPlotChart
                         />
                     ))}
                 {this.points}
-                <VerticalColorLegend manager={this} />
+                <VerticalColorLegendComponent
+                    legend={this.verticalColorLegend}
+                    x={this.legendX}
+                    y={this.legendY}
+                    activeColors={this.activeColors}
+                    focusColors={this.focusColors}
+                    onLegendMouseOver={this.onLegendMouseOver}
+                    onLegendMouseLeave={this.onLegendMouseLeave}
+                    onLegendClick={this.onLegendClick}
+                />
                 {sizeLegend && (
                     <>
                         {separatorLine(ySizeLegend)}
